Implement VeinsVehicleDataProvider as plain C++ helper and verify initialization
